Rework renfxx.h C++ header file
Moved implementation to separate cpp files to help with compilation
times for projects relying on e-antic and created a dedicated forwarding
header that's even faster when that is good enough.

This should work at no runtime performance penalty when compiling with
-flto -fno-semantic-interposition as things can still be inlined.

I tried to unify the naming somewhat. This is of course a matter of
taste that one can bikeshed about forever but imho `get` is a strange
prefix when the "const" basically already setting that this is a reading
operation. Where do you draw the line between `get_num` and
`(get_)floor`?  I tried to keep all the old methods intact and marked
them as depracated.

In some cases I had to change the semantics of methods, i.e., changing
the return value of `renf_elem_class::parent`. This is of course a hard
API change. If this is too extreme we can deprecate parent and call this
method `renf()`/`number_field()` or something like that.

I generally moved to methods return a `*_t`, say `renf_t` instead of a
`renf`. Imho, you want to feed this element into some low-level C
function and that function usually wants the `_t` variant.

I am implementing (work in progress) each operator as a single template
that then switches depending on the passed type with a `constexpr if`.
I only implement the inplace operators, i.e., `+=` but not `+`. Boost
operators does that for me (and I don't have to worry about efficiency
there anymore.)

Also I added move semantics to the constructors and operator=. I am not
sure that this is of too much use at the moment. But it's probably nice
to have this in the API and implement it more efficiently at some point.

Finally, I am moving some implementation details into anonymous
namespaces so they are not exported, such as the static xalloc.