Use cosmic Ubuntu in Travis
Travis is slow in introducing new versions of Ubuntu, so we have to go
the docker way here. This is a bit annoying but we get our version of
arb and C++17 compilers for free. Note that much older versions of
Ubuntu already have C++17 compilers (both clang and gcc-8) but that they
are not the default compilers there.