Replace renf_elem_class.nf with a shared_ptr
as a preparation to serialization with cereal.

This is still incomplete. We should probably remove the constructors and
provide factory functions. Also, it might be convenient to add
renf_elem_class() constructors that take a shared_ptr<renf_class>
instead of the renf_class reference.

Note that the changes to the test cases might seem like a lot of changes
are necessary to adapt existing code. However this is not the case, I'll
add something to the NEWS to explain that later.