Fix insufficient use of thread locking
while the existing scheme of thread-locking was sufficient to forbid
simultaneous refinements, it was not sufficient to make sure that nobody
reads an nf->emb while some other thread is disposing of it.

We could have fixed this by adding locking to all readers of nf->emb but
this seems quite costly in single-threaded code and also very hard to
verify and maintain.

Instead we now opted for a more restrictive version which requires users
to explicitly mark multi-threaded code by locking emb.