Simplify and extend tests for binary operators
this is a preparation for testing floor division.

We also simplify the generator logic. Making the "iterations"
explicit in the generator is not necessary; instead we can just "take"
from the generator.

Originally, we needed the iterations to make sure that the number field
is still around in renf_elem_generator.hpp to renf_elem_clear the last
element generated. Now, we just hold on to a copy of that last number
field so we do not have to synchronize lifetimes anymore.