Silence compiler warnings about unitialized variables
the evaluation code here uses the variable res as a temporary when it is
not aliased. Otherwise, it creates a new temporary. This optimization is
probably there for a reason but not only the latest GCC finds this
confusing and complains about rres being possibly uninitialized. We try
to make this easier to read and get rid of the compiler warning.