Set prefix when configuring
strangely this fixes a lot of cppyy issues on macOS.

cppyy on macOS does not seem to search the paths specified by LDFLAGS
for libraries. On Linux this does seem to happen.

Setting the prefix adds the CONDA_PREFIX/lib to the search path that is
specified in local.h which then means that #pragma cling load actually
works to find these libraries.

On Linux this does not seem to be necessary. In both setups, cppyy.load
works without any of this.