Build documentation with breathe
realistically, standardese is not going to be in a stable state anytime
soon. We change the build to use breathe instead which is much more
popular. While breathe is also struggling to keep up with development,
it seems to be in much better shape than standardes (which I am
maintaining.)